+++ matrix_lib/matr_sub_matr.c
@@ -12,7 +12,7 @@
 
 #include "matrix.h"
 
-void	matr_sub_matr(double **matrix1, double **matrix2, \
+void	matr_sub_matr(float **matrix1, float **matrix2, \
 int num_rows, int num_cols)
 {
 	int i;
