+++ matrix_lib/matrix.h
@@ -18,27 +18,27 @@
 # include <stdio.h>
 # include <unistd.h>
 
-double		**create_matrix(int num_rows, int num_cols);
-double		**matr_mul(double **matrix1, double **matrix2);
-void		matr_mul_by_scalar(double **matrix, \
+float		**create_matrix(int num_rows, int num_cols);
+float		**matr_mul(float **matrix1, float **matrix2);
+void		matr_mul_by_scalar(float **matrix, \
 int num_rows, int num_cols, int scalar);
-void		matr_div_by_scalar(double **matrix, \
+void		matr_div_by_scalar(float **matrix, \
 int num_rows, int num_cols, int scalar);
-void		matr_add_scalar(double **matrix, \
+void		matr_add_scalar(float **matrix, \
 int num_rows, int num_cols, int scalar);
-void		matr_sub_scalar(double **matrix, \
+void		matr_sub_scalar(float **matrix, \
 int num_rows, int num_cols, int scalar);
-void		matr_add_matr(double **matrix1, \
-double **matrix2, int num_rows, int num_cols);
-void		matr_sub_matr(double **matrix1, \
-double **matrix2, int num_rows, int num_cols);
-double		matr_trace(double **matrix, int num_rows, int num_cols);
-double		**matr_transpose(double **matr);
-void		matr_copy(double **src_matrix, \
-double **dst_matrix, int num_rows, int num_cols);
-void		matr_free(double **matrix, int num_rows);
-int			*matr_to_line(double **matrix, int num_rows, int num_cols);
-void		matrix_identity(double **matrix, int num_rows);
-void		matr_mul_void(double *matrix1, double **matrix2);
+void		matr_add_matr(float **matrix1, \
+float **matrix2, int num_rows, int num_cols);
+void		matr_sub_matr(float **matrix1, \
+float **matrix2, int num_rows, int num_cols);
+float		matr_trace(float **matrix, int num_rows, int num_cols);
+float		**matr_transpose(float **matr);
+void		matr_copy(float **src_matrix, \
+float **dst_matrix, int num_rows, int num_cols);
+void		matr_free(float **matrix, int num_rows);
+int			*matr_to_line(float **matrix, int num_rows, int num_cols);
+void		matrix_identity(float **matrix, int num_rows);
+void		matr_mul_void(float *matrix1, float **matrix2);
 
 #endif
